services/quant-strategy/requirements.txt
# Quant Strategy 微服务依赖

# ============================================
# Web 框架
# ============================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0  # Pydantic 配置管理

# ============================================
# 服务注册发现
# ============================================
aiohttp>=3.9.0

# ============================================
# HTTP 客户端（调用 get-stockdata）
# ============================================
httpx>=0.25.0             # 现代异步 HTTP 客户端
requests>=2.31.0          # 同步 HTTP 客户端（备用）

# ============================================
# 数据库驱动
# ============================================
sqlalchemy>=2.0.0         # ORM 框架
aiomysql>=0.2.0           # MySQL 异步驱动
pymysql>=1.1.0            # MySQL 同步驱动
asyncpg>=0.29.0           # PostgreSQL 异步驱动（备用）
clickhouse-driver>=0.2.6  # ClickHouse 驱动（备用）

# ============================================
# 缓存与消息队列
# ============================================
redis>=5.0.0              # Redis 客户端
hiredis>=2.2.0            # Redis 高性能解析器

# ============================================
# 数据处理与量化计算
# ============================================
pandas>=2.0.0             # 数据分析
numpy>=1.24.0             # 数值计算
scipy>=1.11.0             # 科学计算
numba>=0.58.0             # 高性能计算加速
networkx>=3.0             # 图论计算
python-louvain>=0.16      # Louvain 社区发现
leidenalg>=0.10.0         # Leiden 社区发现 (更稳定)
python-igraph>=0.11.0     # igraph 基础库 (Leiden 依赖)
pyarrow>=15.0.0           # 高性能 Parquet 存储
scikit-learn>=1.3.0       # 机器学习（回测）
# ta-lib>=0.4.28          # 技术分析指标 - 需要C库编译，暂时移除
stockstats>=0.6.2         # 股票统计指标

# ============================================
# 时间处理
# ============================================
python-dateutil>=2.8.0    # 日期工具
pytz>=2023.3              # 时区处理
arrow>=1.3.0              # 现代时间库

# ============================================
# 认证与安全
# ============================================
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4    # 密码哈希

# ============================================
# 监控和日志
# ============================================
structlog>=22.0.0         # 结构化日志
prometheus-client>=0.14.0 # Prometheus 指标
colorlog>=6.7.0           # 彩色日志输出

# ============================================
# 配置与工具
# ============================================
pyyaml>=6.0               # YAML 配置
python-dotenv>=1.0.0      # 环境变量管理
tenacity>=8.2.0           # 重试机制

# ============================================
# 开发与测试
# ============================================
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0         # 测试覆盖率
pytest-mock>=3.11.0       # Mock 支持
faker>=20.0.0             # 测试数据生成
# httpx-mock>=0.11.0      # HTTP mock - Python 3.12 不兼容，暂时移除

# ============================================
# 代码质量
# ============================================
black>=23.0.0             # 代码格式化（可选）
flake8>=6.0.0             # 代码检查（可选）
mypy>=1.5.0               # 类型检查（可选）

# Async database drivers
aiosqlite>=0.19.0
aiomysql>=0.2.0
